backend/requirements.txt
# Core dependencies
pyzotero==1.5.18
boto3==1.34.51
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1

# Database and caching
sqlalchemy==2.0.27
redis==5.0.2
alembic==1.13.1

# Document processing
pypdf2==3.0.1
pdfplumber==0.10.3
python-magic==0.4.27
beautifulsoup4==4.12.3

# Async support
aiofiles==23.2.1
httpx==0.26.0
asyncio==3.4.3

# YAML
pyyaml==6.0.1

# Utilities
python-multipart==0.0.9
watchdog==4.0.0
apscheduler==3.10.4
tenacity==8.2.3

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Logging
loguru==0.7.2